Chained calls of decorator functions fixed
Original PR: https://github.com/daipe-ai/daipe-core/pull/15
