+++ src/daipecore/function/input_decorator_function_basic_test/decorator_functions.py
@@ -1,0 +1,10 @@
+from daipecore.function.input_decorator_function import input_decorator_function
+from injecta.container.ContainerInterface import ContainerInterface
+
+
+@input_decorator_function
+def test_input_function():
+    def wrapper(container: ContainerInterface):
+        return 155
+
+    return wrapper
