+++ src/daipecore/function/input_decorator_function_import_test/input_decorator_function_import_test.py
@@ -1,0 +1,3 @@
+from daipecore.function.input_decorator_function_import_test.test_notebook import load_data
+
+assert 165 == load_data(155)
