--- src/daipecore/function/input_decorator_function_test/input_decorator_function_import_test.py
@@ -1,3 +1,0 @@
-from daipecore.function.input_decorator_function_test.nested_functions_stub import load_data
-
-assert 165 == load_data(155)
